feature(utils): Changes in pseudo-random number generation functionality
Class com.lugowoy.helper.utils.generating.GeneratorRandomNumber.java removed. In my opinion, it is better that such functionality is not static (as in this class), but can be created as an object of a certain class.
The class com.lugowoy.helper.utils.RandomNumber.java was created. It contains the same functionality as GeneratorRandomNumber.java, but not static. The documentation is written.