feature(filling) Moving and fixing a functional interface ../filling.array.strings.ReaderCharacters.java
The functional interface was tightly coupled with ../filling.array.strings package and its classes. For this reason, the interface has been moved to the ../utils package and redone in the class. Thus, he will receive less connectivity and greater expandability in the future. Also, now it can be used as a standalone object.