feature(utils) Changes to the construction of class objects
The method of constructing class objects has been changed:
 CreatorString.java
 ReaderCharacters.java
Prior to this, a static method was used for construction (a kind of Singleton, only the object was not one and not static, but different each time), since the constructor had a private access modifier. Now the constructor has a public modifier and objects are created in the usual way.