refactor(./utils.checking) Code refactoring in class CheckerBound
Methods no longer throw exceptions when validation fails. Now a value of type boolean is returned. The functionality has been changed due to the low level of abstraction of the functional methods of this class. This is not a valid thread of execution in this code and the code that calls it.
Constants of indicated messages for exclusion now have public access. So that they can be used in external code.

Declarations (signatures) of methods are changed.
Now they use generic types. This is implemented due to a change in the internal functionality of the methods. Namely, the compareTo() method from the Comparable<T> interface is now used to check values for correctness.
Also, the parameters now use the final modifier and are marked with the annotation @NotNull.