feature(execution) Changing, refactoring and docs code in the OutputExecutionTimeToConsole class
Now the class implements the contract of the interface OutputExecutionTime;

The boolean isOnlyMilliseconds attribute is added that indicates the formatting style of the appearance of the execution time when it is output to the console. Added getter and setter. A detailed description is in the documentation;

Overloaded constructors are created;

In the output(final long milliseconds) method, a check for the positive value of the argument in milliseconds has been added.

Written code documentation.