refactor(models) Refactoring code in the AbstractArray class
The private field 'size' in the AbstractArray class is initialized with the value of the AbstractArray.START_SIZE static variable

The private field 'modCount' in the AbstractArray class is initialized with the value of the static variable AbstractArray.START_MOD_COUNT

Constructor access modifiers changed from public to protected.

Added exception catching to AbstractArray(int capacity) constructor. The LengthOutOfRangeException exception can be thrown when checking the 'capacity' argument and passed into the more general IllegalArgumentException exception.

In the hashCode() method, instead of calling the getter methods, the variables themselves are called to get the values.