refactor(models) Refactoring code in the ../models/arrays package
Refactoring of the code in the classes of the com.lugowoy.helper.models.arrays package. The reason is the innovations described in commit 00077d29306e33c745ccf58129c205221f19f34f;

Removed constants in AbstractArray class:
 - ARRAY_HEADER;
 - LOWER_CAPACITY;
 - UPPER_CAPACITY;

In the equals() method of the AbstractArray class, when accessing the AbstractArray.modCount field, the AbstractArray.getModCount method was used. Changed to directly reference a variable in this class.